refactor: move common magic methods into a new BaseModel class
