+++ idf_build_apps/app.py
@@ -27,7 +27,6 @@ from packaging.version import (
     Version,
 )
 from pydantic import (
-    BaseModel,
     computed_field,
 )
 
@@ -51,6 +50,7 @@ from .manifest.manifest import (
     Manifest,
 )
 from .utils import (
+    BaseModel,
     BuildError,
     files_matches_patterns,
     find_first_match,
@@ -191,37 +191,6 @@ class App(BaseModel):
             self.build_status.value,
             self._build_duration,
         )
-
-    def __lt__(self, other: t.Any) -> bool:
-        if isinstance(other, App):
-            for k in self.model_dump():
-                if getattr(self, k) != getattr(other, k):
-                    return getattr(self, k) < getattr(other, k)
-                else:
-                    continue
-
-        return NotImplemented
-
-    def __eq__(self, other: t.Any) -> bool:
-        if isinstance(other, App):
-            self_dict = self.model_dump()
-            other_dict = other.model_dump()
-
-            return self_dict == other_dict
-
-        return NotImplemented
-
-    def __hash__(self) -> int:
-        hash_list = []
-        for v in self.__dict__.values():
-            if isinstance(v, list):
-                hash_list.append(tuple(v))
-            elif isinstance(v, dict):
-                hash_list.append(tuple(v.items()))
-            else:
-                hash_list.append(v)
-
-        return hash((type(self),) + tuple(hash_list))
 
     @staticmethod
     def _get_sdkconfig_defaults(sdkconfig_defaults_str: t.Optional[str] = None) -> t.List[str]:
